+++ react_frontend/src/assets/css/formitem.css
@@ -1,3 +1,22 @@
 .form_item {
     width: 280px;
 }
+
+.input_button_container {
+    display: flex;
+    align-items: center;
+    gap: 8px;
+}
+.input_field {
+    width: 70%;
+}
+.modal_button {
+    width: 30%;
+}
+
+.search_btn {
+    flex: 1;
+    display: flex;
+    justify-content: flex-end;
+    align-items: flex-end;
+  }
