+++ react_frontend/src/assets/css/searchform.css
@@ -9,27 +9,3 @@
   margin: 30px; /* 양 옆 여백 */
 }
 
-.form_item {
-  width: 280px;
-}
-
-.input_button_container {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-}
-
-.input_field {
-  width: 70%;
-}
-
-.modal_button {
-  width: 30%;
-}
-
-.full_width {
-  flex: 1;
-  display: flex;
-  justify-content: flex-end;
-  align-items: flex-end;
-}
